pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "n8n-workflow-generator-mcp"
version = "1.0.0"
description = "Claude Code powered n8n workflow generator that creates complete workflow JSON files from natural language descriptions"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["n8n", "workflow", "automation", "mcp", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration :: Automation",
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/your-username/n8n-workflow-generator-mcp"
Repository = "https://github.com/your-username/n8n-workflow-generator-mcp"
Issues = "https://github.com/your-username/n8n-workflow-generator-mcp/issues"

[project.scripts]
n8n-mcp-server = "mcp_server.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp-server"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true